[refactor] Sortierung
