+++ src/main/java/github/weichware10/analyse/gui/analyse/DiagramConfigurator.java
@@ -1,4 +1,4 @@
-package github.weichware10.analyse.gui;
+package github.weichware10.analyse.gui.analyse;
 
 import github.weichware10.analyse.config.DiagramConfig;
 import github.weichware10.analyse.enums.AnalyseType;
