+++ src/main/java/github/weichware10/analyse/gui/general/MainMenuBarController.java
@@ -1,4 +1,4 @@
-package github.weichware10.analyse.gui;
+package github.weichware10.analyse.gui.general;
 
 import github.weichware10.analyse.gui.util.AbsSceneController;
 import github.weichware10.analyse.gui.util.Log;
