+++ src/main/resources/github/weichware10/analyse/gui/analyse/TrialSelector.fxml
