+++ src/main/resources/github/weichware10/analyse/gui/general/FunctionChooser.fxml
