+++ src/main/resources/github/weichware10/analyse/gui/general/MainMenuBar.fxml
